Partial implementation of user->post index
This implementation is incomplete and messy.
One issue is that we need cases for if the user has posted before or
not.
This is because if they haven't posted, their key doesn't exist in the
index.
One way around this is to add the user to the index with an empty post
set when they are added.

Another issue is that public keys cannot be used as keys in Dictionaries
as far as I know. This means we would need some sort of string
idenifier which we use in the index. For this, I just added an arbitrary
username (this is not secure, since there is no checking of the
username).